fix(inspector): fix transform->rotation
description
1.set to (x:-45  y:180 z:0) as r1;
2.dispatch MainEditorTransform;
3.rotation change to r2(not equal r1)

reason
This is because euler angles are not stored internally for an entity. When you set Euler angles, internally, Wonder.js sets a quaternion.

So when you call getLocalEulerAngles, the quaternion is converted back to quaternions.
For example, you’re probably getting 180, 0, 180. This is equivalent to 0, 180, 0.

solution
store setted euler angles to editorState;
get euler angles from editorState;